+++ packages/app/src/features/NoteEditor/RichEditor/__tests__/interactions/readonly.dom.test.ts
