+++ packages/app/src/features/NoteEditor/RichEditor/__tests__/spec/formatting.dom.test.ts
